pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "git_gpt"
version = "0.9.0"
authors = [
    {name = "ShinChven", email = "shinchven@gmail.com"},
]
description = "A CLI tool to generate commit messages and issues based on staged Git diffs using OpenAI GPT models or Ollama"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click",
    "openai",
    "gitpython",
    "tomli",
    "requests",
]

[project.urls]
"Homepage" = "https://github.com/ShinChven/git-gpt.git"
"Bug Tracker" = "https://github.com/ShinChven/git-gpt/issues"

[project.scripts]
git-gpt = "git_gpt.main:cli"

[tool.setuptools]
packages = ["git_gpt"]

[tool.setuptools.package-data]
git_gpt = ["*"]

[tool.setuptools.exclude-package-data]
"*" = ["assets"]